+++ IcVibrations/src/IcVibrations.Core/Mapper/PiezoelectricProfiles/Circular/IPiezoelectricCircularProfileMapper.cs
@@ -1,0 +1,9 @@
+﻿using IcVibrations.Common.Profiles;
+
+namespace IcVibrations.Core.Mapper.PiezoelectricProfiles.Circular
+{
+    /// <summary>
+    /// It's responsible to build the piezoelectric circular profile.
+    /// </summary>
+    public interface IPiezoelectricCircularProfileMapper : IPiezoelectricProfileMapper<CircularProfile> { }
+}